full_rag/requirements.txt
sentence-transformers==2.6.1
faiss-cpu==1.9.0.post1
numpy==1.26.4
llama-cpp-python==0.2.90
pandapower==2.14.11
fastapi==0.115.6
uvicorn==0.32.1
